Fix Buildkite issue when check out merge commits
When you are checking out a merge commit and you provide the GitHub
events merge commits, Buildkite kind of ignores the branch and tries to
checkout the specific SHA ignoring that the Git ref Object that this sha
belong to is not the pull/{id}/head but another one.

One solution to solve this would be that the default BK checkout out
perform fetch on the `pull/{id}/*` so that the SHA will work regardless.

Signed-off-by: Alexandros Sapranidis <alexandros@elastic.co>